backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
asyncio
termcolor
google-generativeai
psycopg2-binary
python-dotenv
crewai
requests
